[firebase-config] renames libs/push_notifications.py to push_notification_config.py
