+++ api/study_api.py
@@ -9,7 +9,7 @@ from database.study_models import Study, StudyField
 from database.user_models import Participant, ParticipantFieldValue
 from authentication.admin_authentication import (authenticate_researcher_study_access,
     get_researcher_allowed_studies, get_session_researcher, researcher_is_an_admin)
-from libs.push_notifications import check_firebase_instance
+from libs.push_notification_config import check_firebase_instance
 
 study_api = Blueprint('study_api', __name__)
