+++ api/survey_api.py
@@ -5,7 +5,7 @@ from database.survey_models import Survey
 from authentication.admin_authentication import authenticate_researcher_study_access
 from libs.dev_utils import p
 from libs.json_logic import do_validate_survey
-from libs.push_notifications import (repopulate_absolute_survey_schedule_events,
+from libs.push_notification_config import (repopulate_absolute_survey_schedule_events,
     repopulate_relative_survey_schedule_events, repopulate_weekly_survey_schedule_events)
 
 survey_api = Blueprint('survey_api', __name__)
