+++ pages/system_admin_pages.py
@@ -15,7 +15,7 @@ from database.system_models import FileAsText
 from database.user_models import Researcher, StudyRelation
 from libs.copy_study import copy_existing_study
 from libs.http_utils import checkbox_to_boolean, string_to_int
-from libs.push_notifications import check_firebase_instance, update_firebase_instance
+from libs.push_notification_config import check_firebase_instance, update_firebase_instance
 from pages.message_strings import (ALERT_ANDROID_DELETED_TEXT, ALERT_ANDROID_SUCCESS_TEXT,
     ALERT_ANDROID_VALIDATION_FAILED_TEXT, ALERT_DECODE_ERROR_TEXT, ALERT_EMPTY_TEXT, ALERT_FIREBASE_DELETED_TEXT,
     ALERT_IOS_DELETED_TEXT, ALERT_IOS_SUCCESS_TEXT, ALERT_IOS_VALIDATION_FAILED_TEXT, ALERT_MISC_ERROR_TEXT,
